+++ crates/flux-middle/src/rty/fold.rs
@@ -1020,8 +1020,8 @@ TrivialTypeTraversalImpls! {
     crate::rty::BoundReftKind,
     crate::rty::BvSize,
     crate::rty::KVid,
-    crate::FluxDefId,
-    crate::FluxLocalDefId,
+    crate::def_id::FluxDefId,
+    crate::def_id::FluxLocalDefId,
     rustc_span::Symbol,
     rustc_hir::def_id::DefId,
     rustc_hir::Safety,
